+++ source/file_utils.h
@@ -12,8 +12,8 @@ char *DirGetRunningPath(void);
 char *DirGetBiosFolderPath(void);
 char *DirGetScreenshotFolderPath(void);
 
-void FileLoad_NoError(const char *filename, void **buffer, unsigned int *size_);
-void FileLoad(const char *filename, void **buffer, unsigned int *size_);
+void FileLoad_NoError(const char *filename, void **buffer, size_t *size_);
+void FileLoad(const char *filename, void **buffer, size_t *size_);
 
 int FileExists(const char *filename); // Returns 1 if file exists
